+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 - 从 `index.html` 文件中获得样式文件 `base.*.css` | Get style file `base.*.css` from `index.html` file.
 - 实现拖拽组件 `DragBall` | Implement drag component `DragBall`.
 - 添加 `.eslintrc.cjs`, `.prettierignore` 与 `.prettierrc.toml` 配置文件 | Add `.eslintrc.cjs`, `.prettierignore` and `.prettierrc.toml` config files.
+- 实现搜索抽屉组件 `SearchDrawer` | Implement search drawer component `SearchDrawer`.